HPO_containerized/components/evaluate/runtime-requirements.txt
# Generated by KFP.
cloudml-hypertune==0.1.0.dev6
google-cloud-logging>=3.11.3
pandas>=2.2.3
scikit-learn>=1.5.2